+++ src/automol/__init__.py
@@ -2,9 +2,9 @@
 
 __version__ = "0.0.7"
 
-from . import geom, qc, types
+from . import geom, graph, qc, types
 from .geom import Geometry, geometry_hash
 from .rd import mol
 from .view import View
 
-__all__ = ["geom", "qc", "types", "Geometry", "geometry_hash", "mol", "View"]
+__all__ = ["geom", "graph", "qc", "types", "Geometry", "geometry_hash", "mol", "View"]