+++ tests/test_graph.py
@@ -1,0 +1,10 @@
+"""Graph tests."""
+
+from automol import graph
+
+
+def test__inchi() -> None:
+    """Test graph inchi."""
+    water_inchi = "InChI=1S/H2O/h1H2"
+    water = graph.from_inchi(water_inchi)
+    assert graph.inchi(water) == water_inchi